API Reference

Diary.configure — Method

configure(; kwargs...)

Change configuration for Diary.jl in the current session.

Supported keywords

author: Author name put in the header. (default: "")

date_format: Dates.format–compatible date format used in the header. (default: "E U d HH:MM")

diary_file_name: Default name of the diary file. (default: "diary.jl")

Note

If ENV["JULIA_DIARY"] is set, the diary file name will be ignored.

Diary.find_diary — Function

find_diary(previous_diary_file=nothing)

Locate the diary file. If previous_diary_file is nothing or is different from the located diary file, also set the global write_header flag to true.

Diary.parse_history — Method

parse_history(history_lines)

Parse the lines in history_lines, strip trailing semi-colons, and determine if they should be written to the diary based on the mode in which they were entered.

Diary.start_watching — Method

start_watching(args...)

Create and schedule a watch_task with the given args. See also watch_task.

Diary.watch_task — Function

watch_task(history_file, repl_history_file=nothing)

Start watching the history file at filepath history_file for changes, and parse those changes to update the diary. If repl_history_file is set to a value other than nothing, also copy changes to repl_history_file.

Diary.write_header — Method

write_header(io)

Write a header to the IO stream, io.

Diary.DiaryConfig — Type

DiaryConfig

Mutable configuration for Diary.jl. Diary.jl contains a single global instance of this type, GLOBAL_CONFIG. Configuration of Diary.jl should be done through configure rather than direct manipulations of GLOBAL_CONFIG.

Diary.TaskThunk — Type

thunk = TaskThunk(f, args)

To facilitate precompilation and reduce latency, we avoid creation of anonymous thunks. thunk can be used as an argument in schedule(Task(thunk)). Adapted from Revise.