Developer reference

Diary.start_watching — Function

start_watching(args...)

Create and schedule a watch_task with the given args. See also watch_task.

Diary.watch_task — Function

watch_task(history_file, repl_history_file=nothing)

Start watching the history file at filepath history_file for changes, and parse those changes to update the diary. If repl_history_file is set to a value other than nothing, also copy changes to repl_history_file.

Diary.parse_history — Function

parse_history(history_lines)

Parse the lines in history_lines, strip trailing semi-colons, and determine if they should be written to the diary based on the mode in which they were entered.

Diary.parse_command — Function

parse_command(cmd)

Parse the diary command, cmd. Valid commands are:

commit [n]: Commit the last n segments.

Diary.find_diary — Function

find_diary(; configuration=read_configuration())

Locate the diary file. The default diary name and blacklist is read from configuration. See also read_configuration().

Diary.read_configuration — Function

read_configuration(filename=find_configuration_file())

Read the configuration from filename. See also find_configuration_file.

Diary.find_configuration_file — Function

find_configuration_file()

Locate the configuration file.

Diary.default_configuration — Function

default_configuration()

Return the default configuration.

Diary.commit — Function

commit(n; kwargs...)

Commit the n latest recorded lines to the diary file.

Keyword arguments

configuration: (default: read_configuration())

diary_file: (default: find_diary(; configuration))

with_header: Write header before lines. (default: true)

Diary.write_header — Function

write_header(io)

Write a header to the IO stream, io.

Diary.TaskThunk — Type

thunk = TaskThunk(f, args)

To facilitate precompilation and reduce latency, we avoid creation of anonymous thunks. thunk can be used as an argument in schedule(Task(thunk)). Adapted from Revise.